+++ app/src/main/java/com/poti/android/data/repository/SearchRepositoryImpl.kt
@@ -37,7 +37,7 @@ class SearchRepositoryImpl @Inject constructor(
                         artistId = item.artistId,
                         postImage = item.postImage,
                         postTitle = item.postTitle,
-                        postCount = item.postCount,
+                        postCount = item.postCount.toLong(),
                         tag = item.tag,
                     )
                 },
