+++ app/src/main/java/com/poti/android/domain/model/search/PartySearchResult.kt
@@ -10,6 +10,6 @@ data class PartySearchItem(
     val artistId: Long,
     val postImage: String,
     val postTitle: String,
-    val postCount: Int,
+    val postCount: Long,
     val tag: String?,
 )
